17-05-2025

+++ src/controllers/productController.js
@@ -1,4 +1,5 @@
 let { ProductService } = require("../services/index")
+let { CartService } = require("../services/index")
 
 class ProductController {
     static async FetchAll(req, res){
@@ -11,20 +12,11 @@ class ProductController {
     }
     static async FetchById(req, res){
         try{
-            let id = req.params
+            let { id } = req.params
             let response = await ProductService.FetchById(id)
             res.status(200).json({success: true, data: response})
         } catch(err){
             res.status(404).json({success: false, message: "unable to fetch product"})
-        }
-    }
-    static async FetchByName(req, res){
-        try{
-            let name = req.params
-            let response = await ProductService.FetchByName(name)
-            res.status(200).json({success: true, data: response})
-        } catch (err){
-            res.status(404).json({success: false, message: "unable to fetch product by name"})
         }
     }
     static async AddProduct(req, res){
@@ -38,7 +30,7 @@ class ProductController {
     }
     static async UpdateProduct(req, res){
         try{
-            let id = req.params
+            let { id } = req.params
             let dt = req.body
             let product = await ProductService.FetchById(id)
 
@@ -54,10 +46,18 @@ class ProductController {
     static async DeleteProduct(req, res){
         try{
             let { id } = req.params
+            let product = await CartService.FetchByProductId(id)
+
+            if(product.length > 0){
+                res.status(404).json({success: false, message: "The product Is in cart so unable to delete."})
+                return
+            }
+
             let response = await ProductService.DeleteProduct(id)
             if(response){
                 res.status(200).json({success: true, message: "Product Deleted successfully"})
             }
+
         } catch(err){
             res.status(404).json({success: false, message: "unable to delete product"})
         }
